Rewrite planner script
Summary:
As we prepare to ship distributed builds, I want all the pieces to be as clear and debuggable as possible, and this script was in my opinion quite hard to maintain. I've re-written it and the parts of the bazel logic that depend upon it to be easier to debug.

The primary motivation behind the change is the nature of the "meta file" format. For each input object file or archive, starlark writes 9 lines into a text file to communicate information to this script. Depending on if we are describing an eager object file or an archive member, some of those lines may be blank.

This format was very hard to read, debug and modify. I have replaced it with a more readable and type checked JSON format.

Otherwise I have removed unnecessary parts, refactored for clarity and added comments in places where it may not be obvious what we are doing.

This should be an entirely non-functional change, just easier to maintain.

Reviewed By: rmaz

Differential Revision: D66734636

fbshipit-source-id: 9d473c37a57baa494fc05c1ab3f43afbbeb9f332