Decouple unconfigured target filtering logic between PTT and classic
Summary:
decouple

PTT and Classic used to share the same unconfigured target filtering API -- there's this function `do_filter()` which takes a boolean flag `exclude_targets_with_special_labels` where classic always set that flag to true and ptt always set that flag to false.

While working on D66525151 I realized that if PTT and Classic don't semantically share the same unconfigured target filtering logic, there's no reason why they should invoke the same API for that, especially given that `do_filter()` will "un-share" the underlying logic anyway by branching on `the `exclude_targets_with_special_labels` flag `.

This diff rewrites the unconfigured target filtering logic on the PTT side so it only minimally depend on what we do in classic -- the only shared thing between the two now is that we want to recognize the same kinds of python targets. Beyond that the code path are different, which lets me re-structure the logic in a way that fits better with the surrounding PTT code.

Reviewed By: connernilsen

Differential Revision: D66550825

fbshipit-source-id: 396a8eb07b971616f882d61f565b11ea655ec9f1