Add separate debug info flag for libraries
Summary:
When we declare artifact in buck2 we *have to* provide it. There is no "optional artifact" api. That means if during compilation we need to declare two outputs: `.o` and `.dwo` file, *both should be present*. `.dwo` file existance is controlled by `-gsplit-dwarf=split` compiler flag.
So in prelude we have to know in advance either we have `.dwo` or not. We have several options to know:

### "Hacky" approach
We can derive file existence by checking array of compiler flags for `gsplit-dwarf=split` entry. That is fragile. Flags can be declared/concatenated by user however user wants.
So relying of compiler flag is slow ticking bomb.

### Ideal world

Ideally this could be represented as enum with associated value, something like
```
enum DebugInfo {
    Separate(split_dwarf_flags),
    Merged,
}
```
And then we'd declare artefact and add flags based on `Separate` case. Unfortunately

### "Well, ok" implementation
So I chose approach with providing two values in macro: setting compiler flag and `separate_debug_info` boolean separatedly. That's not ideal because it brings invariants, but I do not see any better approach.

Reviewed By: milend

Differential Revision: D60671904

fbshipit-source-id: daf7e4b9a10e1e1d82e855b3766cdfac9aa2ca89