Pass post build parameters via protobuf
Summary:
Instead of using command line argument, we want to pass the post build parameters to kotlincd/javacd via protobuf.
It is easier to visualize the parameters this way and helps in debugging.
As next step we can remove redundant parameters from the message and reuse the ones that already exist in build command.

More context: T210437469

Differential Revision: D67598556

fbshipit-source-id: bdc221f447fb80bbbf436d660ce85d2312092227