Split Buck targets in generation to ones represented in Xcode and ones not
Summary:
Not every Buck target is going to have an Xcode equivalent. Nor do they need a target in Xcode, we just want there files included.

For example, `apple_resource` targets don't need a target in Xcode and don't need to be built. We just need to include their files.

To achieve this, we can split our supported rule types into:
1. rule types with xcode targets and and ones without
2. for ones without, collect their files and include them as part of the project target

In this diff, we do this for `apple_resource`, which results in:
1. the files being added to target info for HelloWorldBundle
2. removal of the unnecessary target in the generated workspace

Reviewed By: perehonchuk

Differential Revision: D67841935

fbshipit-source-id: 372379288dbe423945714126cdbdddf2c0032819