Do not collect all linkables we link dynamically
Summary:
#buildmore

This diff changes how we structure dynamic libraries and changes linkables in argfile of executable.

Before that diff, linkables of following dependency graph:
```
:bin → :libA → :libB
```
Without link groups linkables of `:bin` are:
* `bin` srcs
* `libA-for-interface.so`

With link groups before that diff linkables are:
* `bin` srcs
* `libA-for-interface.so`
* `libB-for-interface.so` ← this one shouldn't be here

After that diff linkables of `:bin` with link groups are the same as without link groups:
* `bin` srcs
* `libA-for-interface.so`

## Impact on static link (`mode/opt`)
Should be noop for "static" link e.g. `mode/opt`. Because `get_lib_output_style` returns `static_pic` for `static` link strategy and it's logic matches with what we have before that diff.

Differential Revision: D67302526

fbshipit-source-id: 3f547661aff050837695d90d1d4a93ae225c25fa