Disable link order fixup until the perf fix is ready
Summary:
The change in D67197102 makes the link order in link groups consistent with the _regular_ link order. However, the implementation has a perf and memory impact that affects e.g. Glean indexing.

The perf fix is underway but until it's ready we should temporarily disable the link order fixup.

#buildall

Reviewed By: patskovn, iamirzhan

Differential Revision: D68074196

fbshipit-source-id: e04cbde765db3e5c7342dc430c2e67e875e1be98