Selective Debugging: fix bug with SelectedDebugPaths.txt containing unselected/unfocused targets
Summary:
**Problem**

The `SelectedDebugPaths.txt` contains a list of targets which are not actually focused due to how Swift files are treated. Any targets which include Swift debug info would end up being in `SelectedDebugPaths.txt`, even if not selected/focused.

**Solution**

Ensure that `SelectedDebugPaths.txt` is only computed using the targets which are selected, excluding any targets for debug info purposes.

Reviewed By: chatura-atapattu, manicaesar

Differential Revision: D68146690

fbshipit-source-id: 343adc398bface18b6de91598686e0097dee9702