Always create .dwp from .so or executable, not from .dwo directly
Summary:
We don't actually *need* to create `.dwp` file from list of `.dwo` files in unpacked split dwarf. It just a possiblity, not a requirement. Initially I mistakenly thought we have an option to use `.so` as a `.dwp` root only when we have packed split dwarf.

## Why generating `.dwp` from `.so` is better than from list of `.dwo` files

* First of all, that is more bullet-proof. After trying to implement latter I faced a lot of "duplicated DWO ID" challenges that do not exist in generating from `.so`, because:
* Linker deduplicates symbols in `.so` files, that eliminates duplications and makes resulting `.dwp` files *significantly* smaller in size (~30% smaller) due to less stuff to index after deduplication
* It is simply better to have one paved path if other does not bring significant benefits

## How does it work

When split dwarf is disabled, all debug info is stored in `.o` file and then *copied* into resulting `.so` files (or binary). When split dwarf enabled we have two options: packed and unpacked split dwarf.
With packed split dwarf we still have debug info stored in `.o` file. But it is not copied `so` files, libraries reference `.o` file instead:

```
$ readelf --wide --debug-dump=str-offsets [...]/testinfra/playground/cpp/__example__/libtestinfra_playground_cpp_example.so

Contents of the .debug_str_offsets section:

    Length: 0xc
    Version: 0x5
       Index   Offset [String]
           0       19 .
           1       7a [...]/testinfra/playground/cpp/__example__/__objects__/Example.cpp.pic.o
```
When we enable unpacked split dwarf, *`.so` file still references compile option*, but just references `.dwp` instead of `.o` file:

```
# Same library, but with unpacked dwarf

Contents of the .debug_str_offsets section:

    Length: 0xc
    Version: 0x5
       Index   Offset [String]
           0       19 .
           1       7a [...]/testinfra/playground/cpp/__example__/__objects__/Example.cpp.pic.dwo
```
Notice how the extension is changed.

So we can use `.so` file as an entrypoint to `llvm-dwp` for creating `.dwp` files both in packed and unpacked split dwarf.

### Total amount of possible options

1. [packed-no-dwp] Debug info in `.o` file, .so references it → debuger parses references to `.o` from `.so` and works
2. [packed-with-dwp] Debug info in `.o` file, .so references it, .dwp references it → debugger parses refereces to `.o` from `.dwp` (finding it by convention) and works
3. [unpacked-no-dwp] Debug info in `.dwo` file, `.so` references it → debugger parses references to `.dwo` from `.so` and works
4. [unpacked-with-dwp] Debug info in `.dwp` file, there are NO references to it → debugger parses contents of `.dwp` file (finding it by convention) and works

Reviewed By: asm89

Differential Revision: D68817857

fbshipit-source-id: f7fe79cb51f3c78b407945a6c971503f96e60333