Add local_linker_script_flags attribute to prebuilt_cxx_library
Summary: Some `prebuilt_cxx_binary` need to pass in linker scripts if link groups is not used. Link group attributes are not typed right now, so we can't pass in a linker script file as an arg. Add an extra field `local_linker_script_flags` that's only read when not using link groups, this will be removed once link group map migrates to rules.

Reviewed By: milend

Differential Revision: D68723768

fbshipit-source-id: 6980e44dcfd8c5672b109e1634be3720b05f9b22