Don't pass srcs as additional_classpath_entries in kotlin_test
Summary: The current behaviour is wrong (the second parameter to `build_kotlin_library` is `additional_classpath_entries`), but in a benign way (we just have a bunch of stuff on the classpath that we can't actually read or use). However it breaks when the signature changes in the following diff, so fixing it here.

Reviewed By: navidqar

Differential Revision: D69120904

fbshipit-source-id: 96d1bc057cd4e5f88a1ea6a615b2b6f421cf9ea6