Fix JavaCompilingDepsTSet so that we don't create multiple nodes with
Summary:
TSets are deduplicated by _node_, not by _value_. Unfortunately, when we're creating TSets to be used in the `JavaGlobalCodeInfo` map we create different TSets that have the same value (but different children). That means that we end up with the values being duplicated, which we don't want.

Instead, we make it so that we only create one TSet per value, and then if we want to reuse that value, we just create a new TSet that has it as a child rather than setting the value on it again.

Is this `JavaGlobalCodeInfo` a confusing mess? Yes, I think it is.

Reviewed By: navidqar

Differential Revision: D69051758

fbshipit-source-id: 4dc3f7edf85418e54688832eb77d2b417dbd137b