Fix the order and declaration of property sheets in VSGO generated projects/mode configs
Summary:
Prior to this change, the property sheet includes weren't in a "PropertySheets" labeled property group so they were not correctly displayed in the Property Manager.
Additionally, the order of specifying the property sheets was nonstandard, resulting in some settings (most notably C++ language standard) not being picked up correctly.

Reviewed By: autozimu

Differential Revision: D69069124

fbshipit-source-id: 9e3ba7c4438dd5a68530d8995facdf4c567c4ffd