Allow overriding directory name for filegroup
Summary: Filegroup rules are often useful in macros for rearranging artifacts before passing them to a downstream rule, e.g. arranging the structure of sources to be passed to a `zip_file` rule. Filegroups output a directory with the name of the rule, which can get awkward when using them in a macro this way. This commit allows specifying a custom directory name so a filegroup rule in a macro can be given a non-conflicting name without affecting the output.

Reviewed By: IanChilds

Differential Revision: D68931519

fbshipit-source-id: e056996ead436a9bf2fc334bed732020488c0d8e