Add option to use thin archives and --start-lib --end-lib style linker invocations
Summary:
When linking an executable, each apple_library depended upon is usually packaged as a static archive. While we want archive linking semantics, we often don't want the archive itself.

We may instead pass the archive members to the linker between `--start-lib` and `--end-lib` flags without ever creating the archive. The hope is that by avoiding building the archive at all, we may speed up builds and save on RE resources.

For the time being, this change excludes distributed thin-LTO, but that will come as a follow up.

Reviewed By: rmaz

Differential Revision: D68649050

fbshipit-source-id: 33497fc23a1a04668ff9dae9338362d0f1639824