Add option to use thin archives and --start-lib --end-lib style linker invocations (attempt 2)
Summary:
When linking an executable, each apple_library depended upon is usually packaged as a static archive. While we want archive linking semantics, we often don't want the archive itself.
We may instead pass the archive members to the linker between --start-lib and --end-lib flags without ever creating the archive. The hope is that by avoiding building the archive at all, we may speed up builds and save on RE resources.
For the time being, this change excludes distributed thin-LTO, but that will come as a follow up.

This change was reverted a couple times, but only because changes to the linker invocation crafting logic this depends upon caused issues. This is a non functional change on its own.

Reviewed By: rmaz

Differential Revision: D69948536

fbshipit-source-id: 41a41dfd482ec316bc7baad7db09375cf9db9295