Create cuda_compile_style Buck constraints and attribute
Summary:
We use NVCC to compile CUDA code, and NVCC is a compiler driver that calls multiple other tools when it sees a CUDA file. This diff introduces the cuda_compile_style cfg modifier, which has two options:
* Mono - this refers to the current way of how we use NVCC, which is calling a single NVCC command on a CUDA file and executing that along with its sub-commands in a single Buck action.
* Dist - instead of letting NVCC drive the CUDA compilation, we pass `-dryrun` to NVCC to let it print out the commands it plans to run, parse those commands, and run them in separate Buck actions.

The commands that NVCC runs for a single CUDA file can be split into the host code path and the device code path:

# Host code path
* Call the C++ preprocessor to generate a .cpp4.ii IR code.
* Call a close-sourced tool called cudafe++ which processes the .cpp4.ii file into a .cudafe1.cpp file.
* Call the C++ compiler to compile the .cudafe1.cpp file into an object file.

# Device code path
For each codegen target:
* Call the C++ preprocessor to generate a .cpp1.ii IR code.
* Call a close-sourced tool called cicc, which is probably an optimizer, to lower the code into .ptx code. PTX is a lower-level IR of CUDA.
* PTX can be included in the object file directly for JIT-compilation at runtime, but usually, NVCC then calls ptxas, which is the PTX assembler, to lower the code further into .cubin format.
* Call another close-sourced tool called fatbinary to bundle all lowered PTX and cubin code into .fatbin.c. .fatbin.c is included into .cudafe1.cpp.

(If separate compilation is enabled, where device code can reference device code from another file, a couple device linking steps follow, which depend on the nvlink and fatbinary tools. NCCL uses separate compilation).

See the [NVCC doc](https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html) for more details on the CUDA compilation process.

In fbcode, we by default enable both the A100 and H100 targets (and soon B200 target), and we codegen the PTX and cubin code for all targets. The codegen is also configurable using the `fbcode.nvcc_arch` buck config, where some fbcode targets may choose to only target a single GPU architecture.

Due to the fact that the entire NVCC compilation pipeline is defined as a single Buck action, these compilation sub-commands don't get to share cache. For example, compiling a CUDA file for A100,  and then compiling the same CUDA file for all targets, do not share cache with each other. By breaking up the CUDA compilation pipeline into individual actions, we aim to increase cache hit rate and improve CUDA build speed.

This diff creates the `cuda_compile_style` cfg modifier, selects on it in `cpp_common.bzl`, and passes it down to the Buck rules. As of this diff the attribute is not used and will be used in later diffs in this stack.

Reviewed By: scottcao

Differential Revision: D69757710

fbshipit-source-id: f88dca7ecd21d33a429227178d81c43fa3856f4d